Forward Inference works!
So I essentially take the function and "optimize" it by
rewriting it to a standard format which takes in the
varying parameters as a tuple of Float32's and
the globals as an array of Float32's. The fixed parameters
are parsed away and replaced in the function body
by their values.